+++ frontend/scripts/react-components/data-portal/data-portal-disabled-message.component.jsx
