+++ frontend/scripts/react-components/data-portal/data-portal-page.container.js
@@ -1,7 +1,8 @@
+import { bindActionCreators } from 'redux';
 import { loadContextNodes } from 'actions/data.actions';
 import { trackDataDownloadFormLoaded, trackDownload } from 'analytics/analytics.actions';
 import { connect } from 'react-redux';
-import DataPortalPage from 'react-components/data/data-portal-page.component';
+import DataPortalPage from 'react-components/data-portal/data-portal-page.component';
 
 const mapStateToProps = state => ({
   contexts: state.data.contexts,
@@ -10,17 +11,11 @@ const mapStateToProps = state => ({
   indicators: state.data.indicators
 });
 
-const mapDispatchToProps = dispatch => ({
-  onContextSelected: (contextId) => {
-    dispatch(loadContextNodes(contextId));
-  },
-  onDownloadTriggered: (params) => {
-    dispatch(trackDownload(params));
-  },
-  onDataDownloadFormLoaded: () => {
-    dispatch(trackDataDownloadFormLoaded());
-  }
-});
+const mapDispatchToProps = dispatch => bindActionCreators({
+  onContextSelected: loadContextNodes,
+  onDownloadTriggered: trackDownload,
+  onDataDownloadFormLoaded: trackDataDownloadFormLoaded
+}, dispatch);
 
 export default connect(
   mapStateToProps,
