+++ frontend/scripts/react-components/data-portal/data-portal.thunks.js
@@ -1,0 +1,3 @@
+import { loadContext } from 'actions/data.actions';
+
+export const getDataPortalContext = dispatch => dispatch(loadContext());
